Implement data retrieving logic from remote DB [WIP] Complete the remaining JSON parsing logic

+++ src/java/model/Location.java
@@ -1,25 +1,25 @@
 package model;
 
 public class Location {
-    private final int x;
-    private final int y;
-    private final int z;
+    private final double x;
+    private final double y;
+    private final double z;
 
-    public Location(int x, int y, int z) {
+    public Location(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
